+++ createGraph.js
@@ -1,0 +1,18 @@
+const canvas = document.getElementById('grafico')
+
+let grafico;
+export const crearGrafico = (data) => {
+    if (grafico) {
+        grafico.destroy()
+    }
+    grafico = new Chart(canvas, {
+        type: 'bar',
+        data: {
+            labels: data.fechas,
+            datasets: [{
+                label: data.type,
+                data: data.info
+            },]
+        }
+    })
+}
