+++ index.html
@@ -22,5 +22,5 @@
 </body>
 <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-<script src="app.js"></script>
+<script src="app.js" type="module"></script>
 </html>
