+++ package.json
@@ -1,0 +1,13 @@
+{
+  "name": "proyecto_dashboard",
+  "version": "1.0.0",
+  "description": "--- ## Datos COVID-19",
+  "main": "app.js",
+  "type": "module",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
